split subscription page into tabs and add new items in tabs'
