+++ client/AdminUI/src/app/components/subscriptions/manage-subscription/subscription-config-tab/subscription-config-tab.component.scss
@@ -1,0 +1,16 @@
+.customer-card {
+    font-size: 16px !important;
+    font-weight: bold;
+  }
+  
+  .customer-card .placeholder {
+      font-weight: normal;
+      color: #cccccc;
+  }
+  
+  .startButton {
+      align-self: right;
+  }
+  .active_subscription {
+      margin-top: 1em
+  }
