+++ client/AdminUI/src/app/components/subscriptions/manage-subscription/subscription-stats-tab/subscription-stats-tab.component.scss
@@ -1,0 +1,8 @@
+.cycle-select-dropdown {
+    min-width:25em !important;
+}
+
+.cycle-select-wrapper {
+    display: inline-block;
+    margin-left: 1em;
+}
