+++ client/AdminUI/src/app/components/subscriptions/sub-dashboard/sub-dashboard.component.html
@@ -1,9 +1,9 @@
-<mat-card class="mb-4">
-  <mat-card-header>
+<mat-card *ngIf="dataAvailable" class="mb-4">  
+  <!-- <mat-card-header>
     <mat-card-title style="margin-left: -16px !important">
       Statistics For Current Cycle
     </mat-card-title>
-  </mat-card-header>
+  </mat-card-header> -->
   <mat-card-content>
     <div class="d-flex flex-column">
       <div style="max-height: 200px; margin-bottom: 1rem;">
