+++ controller/src/api/urls.py
@@ -33,6 +33,7 @@ schema_view = get_schema_view(
     ),
     public=True,
     permission_classes=(permissions.AllowAny,),
+    url="http://localhost:8000"
 )
 
 urlpatterns = [
